refactor(nlu): extract intent data models into separate module
Separated pure data models (FactoidCreateIntent, FactoidQueryIntent,
ModificationType) from detection logic for better modularity.

- Created intents.py with frozen dataclasses for all intent types
- Added missing FactoidQueryIntent (needed by question_parser)
- Updated factoid_intent_detector to import from intents module
- Fixed question_parser imports (was importing non-existent module)
- Fixed nlu/__init__.py exports (removed non-existent QuestionIntent)
- Applied modern Python type hints (X | None instead of Optional[X])

All 15 NLU tests pass. Fixes broken test imports.
