+++ .beads/issues.jsonl
@@ -5,7 +5,7 @@
 {"id":"bd-28c","title":"Project structure and package layout","description":"Set up the Python package structure: src/infobot/ with subpackages for db, kb, nlu, formatter, services. Create __init__.py files, update pyproject.toml with runtime dependencies (openai, discord.py, aiosqlite). Establish the skeleton that all other work builds on.","status":"closed","priority":1,"issue_type":"task","created_at":"2026-02-06T05:38:48.461185518Z","created_by":"ubuntu","updated_at":"2026-02-06T06:12:07.544374270Z","closed_at":"2026-02-06T06:12:07.544345817Z","close_reason":"Package structure created, deps installed, imports verified","source_repo":".","compaction_level":0,"original_size":0}
 {"id":"bd-2bt","title":"Legacy factoid import tool","description":"Phase 1: CLI tool to import legacy Infobot factoid files (botname-is.txt, botname-are.txt). Parse 'topic => information' format, clean IRC formatting, apply quality heuristics, import to SQLite knowledge base. Per docs/legacy_import.md spec.","status":"open","priority":2,"issue_type":"feature","created_at":"2026-02-06T05:39:25.806474139Z","created_by":"ubuntu","updated_at":"2026-02-06T05:39:40.048107763Z","source_repo":".","compaction_level":0,"original_size":0,"dependencies":[{"issue_id":"bd-2bt","depends_on_id":"bd-6lc","type":"blocks","created_at":"2026-02-06T05:39:40.048066651Z","created_by":"ubuntu","metadata":"{}","thread_id":""}]}
 {"id":"bd-2us","title":"Modal deployment skeleton","description":"Create modal.py with: Modal app definition, vLLM model serving on T4 GPU, Volume for SQLite persistence, Discord bot as a long-running Modal function. Support modal serve for dev and modal deploy for production. Keep warm configuration.","status":"open","priority":1,"issue_type":"task","created_at":"2026-02-06T05:39:22.969780997Z","created_by":"ubuntu","updated_at":"2026-02-06T05:39:40.023226350Z","source_repo":".","compaction_level":0,"original_size":0,"dependencies":[{"issue_id":"bd-2us","depends_on_id":"bd-18i","type":"blocks","created_at":"2026-02-06T05:39:39.997558683Z","created_by":"ubuntu","metadata":"{}","thread_id":""},{"issue_id":"bd-2us","depends_on_id":"bd-1o7","type":"blocks","created_at":"2026-02-06T05:39:40.023184076Z","created_by":"ubuntu","metadata":"{}","thread_id":""}]}
-{"id":"bd-6lc","title":"Factoid data model and storage","description":"Phase 1 core: Implement factoid CRUD operations. Factoids have a key, value, type (is/are), support for special tags (<reply>, <action>), pipe-separated random responses, and variable substitution ($who, $date). Store and retrieve from SQLite.","status":"open","priority":1,"issue_type":"feature","created_at":"2026-02-06T05:39:04.091764143Z","created_by":"ubuntu","updated_at":"2026-02-06T06:43:39.053689899Z","close_reason":"Completed: factoid data model and CRUD storage with 34 passing tests","source_repo":".","compaction_level":0,"original_size":0,"dependencies":[{"issue_id":"bd-6lc","depends_on_id":"bd-vfs","type":"blocks","created_at":"2026-02-06T05:39:39.742277136Z","created_by":"ubuntu","metadata":"{}","thread_id":""}]}
+{"id":"bd-6lc","title":"Factoid data model and storage","description":"Phase 1 core: Implement factoid CRUD operations. Factoids have a key, value, type (is/are), support for special tags (<reply>, <action>), pipe-separated random responses, and variable substitution ($who, $date). Store and retrieve from SQLite.","status":"closed","priority":1,"issue_type":"feature","created_at":"2026-02-06T05:39:04.091764143Z","created_by":"ubuntu","updated_at":"2026-02-06T20:52:05.717007489Z","closed_at":"2026-02-06T20:52:05.716969088Z","close_reason":"Verified complete: data model, CRUD store, SQLite schema, 34 tests passing","source_repo":".","compaction_level":0,"original_size":0,"dependencies":[{"issue_id":"bd-6lc","depends_on_id":"bd-vfs","type":"blocks","created_at":"2026-02-06T05:39:39.742277136Z","created_by":"ubuntu","metadata":"{}","thread_id":""}]}
 {"id":"bd-gca","title":"Message handler (orchestrator)","description":"Phase 1 core: The main message processing pipeline. Receives a message, runs it through NLU (question parser + intent detector), queries/updates knowledge base, formats response, optionally enhances via LLM. Returns the response to send back to the chat platform.","status":"open","priority":1,"issue_type":"feature","created_at":"2026-02-06T05:39:16.162027179Z","created_by":"ubuntu","updated_at":"2026-02-06T05:39:39.944743933Z","source_repo":".","compaction_level":0,"original_size":0,"dependencies":[{"issue_id":"bd-gca","depends_on_id":"bd-18i","type":"blocks","created_at":"2026-02-06T05:39:39.944700262Z","created_by":"ubuntu","metadata":"{}","thread_id":""},{"issue_id":"bd-gca","depends_on_id":"bd-1cg","type":"blocks","created_at":"2026-02-06T05:39:39.919231866Z","created_by":"ubuntu","metadata":"{}","thread_id":""},{"issue_id":"bd-gca","depends_on_id":"bd-6lc","type":"blocks","created_at":"2026-02-06T05:39:39.841750291Z","created_by":"ubuntu","metadata":"{}","thread_id":""},{"issue_id":"bd-gca","depends_on_id":"bd-qrv","type":"blocks","created_at":"2026-02-06T05:39:39.869311086Z","created_by":"ubuntu","metadata":"{}","thread_id":""},{"issue_id":"bd-gca","depends_on_id":"bd-u88","type":"blocks","created_at":"2026-02-06T05:39:39.894655776Z","created_by":"ubuntu","metadata":"{}","thread_id":""}]}
 {"id":"bd-qrv","title":"Question parser (NLU)","description":"Phase 1 core: Parse incoming messages to detect question intent. Handle formats: 'What is X?', 'Where is X?', 'Who is X?', 'X?', 'tell me about X'. Extract the factoid key being queried. Return structured intent objects.","status":"closed","priority":1,"issue_type":"feature","created_at":"2026-02-06T05:39:06.445401181Z","created_by":"ubuntu","updated_at":"2026-02-06T06:31:24.787214764Z","closed_at":"2026-02-06T06:31:24.787183573Z","close_reason":"Completed","source_repo":".","compaction_level":0,"original_size":0,"dependencies":[{"issue_id":"bd-qrv","depends_on_id":"bd-28c","type":"blocks","created_at":"2026-02-06T05:39:39.766342015Z","created_by":"ubuntu","metadata":"{}","thread_id":""}]}
 {"id":"bd-u88","title":"Factoid creation intent detector","description":"Phase 1 core: Detect when a user is stating a fact ('X is Y', 'X are Y'). Distinguish from questions and commands. Handle edge cases like 'no, X is Y' (replace) and 'X is also Y' (append). Return structured intent with key, value, and modification type.","status":"closed","priority":1,"issue_type":"feature","created_at":"2026-02-06T05:39:09.163270240Z","created_by":"ubuntu","updated_at":"2026-02-06T06:49:31.816140986Z","closed_at":"2026-02-06T06:49:31.816115672Z","close_reason":"Completed","source_repo":".","compaction_level":0,"original_size":0,"dependencies":[{"issue_id":"bd-u88","depends_on_id":"bd-28c","type":"blocks","created_at":"2026-02-06T05:39:39.790453450Z","created_by":"ubuntu","metadata":"{}","thread_id":""}]}
