+++ src/infobot/nlu/__init__.py
@@ -1,16 +1,17 @@
 """Natural language understanding for intent detection and question parsing."""
 
-from infobot.nlu.factoid_intent_detector import (
+from infobot.nlu.factoid_intent_detector import detect_factoid_create_intent
+from infobot.nlu.intents import (
     FactoidCreateIntent,
+    FactoidQueryIntent,
     ModificationType,
-    detect_factoid_create_intent,
 )
-from infobot.nlu.question_parser import QuestionIntent, parse_question
+from infobot.nlu.question_parser import parse_question
 
 __all__ = [
     "FactoidCreateIntent",
+    "FactoidQueryIntent",
     "ModificationType",
-    "QuestionIntent",
     "detect_factoid_create_intent",
     "parse_question",
 ]
